Refactor TSDB test helper structs/functions into separate file
