+++ tsdb/head_test.go
@@ -5981,45 +5981,3 @@ func TestHeadAppender_AppendCTZeroSample(t *testing.T) {
 		require.Equal(t, chunkenc.ValNone, it.Next())
 	}
 }
-
-type sampleTypeScenario struct {
-	sampleType string
-	appendFunc func(appender storage.Appender, lbls labels.Labels, ts int64, value int64) (storage.SeriesRef, error, sample)
-	sampleFunc func(ts, value int64) sample
-}
-
-var sampleTypeScenarios = map[string]sampleTypeScenario{
-	"float": {
-		sampleType: sampleMetricTypeFloat,
-		appendFunc: func(appender storage.Appender, lbls labels.Labels, ts int64, value int64) (storage.SeriesRef, error, sample) {
-			s := sample{t: ts, f: float64(value)}
-			ref, err := appender.Append(0, lbls, ts, s.f)
-			return ref, err, s
-		},
-		sampleFunc: func(ts, value int64) sample {
-			return sample{t: ts, f: float64(value)}
-		},
-	},
-	"integer histogram": {
-		sampleType: sampleMetricTypeHistogram,
-		appendFunc: func(appender storage.Appender, lbls labels.Labels, ts int64, value int64) (storage.SeriesRef, error, sample) {
-			s := sample{t: ts, h: tsdbutil.GenerateTestHistogram(int(value))}
-			ref, err := appender.AppendHistogram(0, lbls, ts, s.h, nil)
-			return ref, err, s
-		},
-		sampleFunc: func(ts, value int64) sample {
-			return sample{t: ts, h: tsdbutil.GenerateTestHistogram(int(value))}
-		},
-	},
-	"float histogram": {
-		sampleType: sampleMetricTypeHistogram,
-		appendFunc: func(appender storage.Appender, lbls labels.Labels, ts int64, value int64) (storage.SeriesRef, error, sample) {
-			s := sample{t: ts, fh: tsdbutil.GenerateTestFloatHistogram(int(value))}
-			ref, err := appender.AppendHistogram(0, lbls, ts, nil, s.fh)
-			return ref, err, s
-		},
-		sampleFunc: func(ts, value int64) sample {
-			return sample{t: ts, fh: tsdbutil.GenerateTestFloatHistogram(int(value))}
-		},
-	},
-}
